+++ agency-service/src/main/java/com/foliageh/service/AgencyService.java
@@ -27,8 +27,8 @@ public class AgencyService {
         return flats.stream().findFirst();
     }
 
-    public Optional<Flat> getMostExpensiveAmongThree(Long id1, Long id2, Long id3) {
-        List<Long> ids = Arrays.asList(id1, id2, id3);
+    public Optional<Flat> getMostExpensiveAmongThree(Integer id1, Integer id2, Integer id3) {
+        List<Integer> ids = Arrays.asList(id1, id2, id3);
         List<Flat> flats = flatRepository.findByIdIn(ids);
         return flats.stream().max(Comparator.comparingDouble(Flat::getPrice));
     }
